update header & renaming
